Split middleware-common tests into separate subcrate
