+++ lib/dev-utils/Cargo.toml
@@ -6,6 +6,7 @@ license = "MIT"
 authors = ["The Wasmer Engineering Team <engineering@wasmer.io>"]
 edition = "2018"
 repository = "https://github.com/wasmerio/wasmer"
+publish = false
 
 [dependencies]
 libc = "0.2.60"
