+++ lib/middleware-common-tests/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "wasmer-middleware-common-tests"
+version = "0.6.0"
+authors = ["The Wasmer Engineering Team <engineering@wasmer.io>"]
+edition = "2018"
+repository = "https://github.com/wasmerio/wasmer"
+license = "MIT"
+publish = false
+
+[dependencies]
+wasmer-runtime-core = { path = "../runtime-core", version = "0.6.0" }
+wasmer-middleware-common = { path = "../middleware-common", version = "0.6.0" }
+wasmer-clif-backend = { path = "../clif-backend", version = "0.6.0" }
+wasmer-llvm-backend = { path = "../llvm-backend", version = "0.6.0", optional = true }
+wasmer-singlepass-backend = { path = "../singlepass-backend", version = "0.6.0", optional = true }
+
+[features]
+clif = []
+llvm = ["wasmer-llvm-backend"]
+singlepass = ["wasmer-singlepass-backend"]
+
+[dev-dependencies]
+wabt = "0.9.1"
+criterion = "0.2"
+
+[[bench]]
+name = "metering_benchmark"
+harness = false
