+++ lib/middleware-common-tests/benches/metering_benchmark.rs
