+++ lib/middleware-common/Cargo.toml
@@ -9,19 +9,3 @@ edition = "2018"
 
 [dependencies]
 wasmer-runtime-core = { path = "../runtime-core" }
-wasmer-clif-backend = { path = "../clif-backend", version = "0.6.0" }
-wasmer-llvm-backend = { path = "../llvm-backend", version = "0.6.0", optional = true }
-wasmer-singlepass-backend = { path = "../singlepass-backend", version = "0.6.0", optional = true }
-
-[dev-dependencies]
-wabt = "0.9.1"
-criterion = "0.2"
-
-[features]
-clif = []
-llvm = ["wasmer-llvm-backend"]
-singlepass = ["wasmer-singlepass-backend"]
-
-[[bench]]
-name = "metering_benchmark"
-harness = false
